rename "square lose turn" to "lose turn"
