+++ ProjectJava/src/main/resources/JsonFiles/fussy.json
@@ -3,56 +3,56 @@
 "GameStart": "rollDiceToEnter",
 "Interaction":{"InteractType":"MoveOtherPlayer","sendPlayerTo":0},
 "Board": {
-	"boardType": {"type":"Loop","rounds":500},
+	"boardType": {"type":"Classic"},
 	"squares": [
 		{"SquareInfo": "Empty"},
 		{"SquareInfo": "Empty"},
-		{"SquareInfo": "SquareLoseTurn"},
+		{"SquareInfo": "loseTurn"},
 		{"SquareInfo": "Empty"},
 		{"SquareInfo": "GetImmunity"},
 		{"SquareInfo": "Empty"},
 		{"SquareInfo": "Empty"},
-		{"SquareInfo": "SquareLoseTurn"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "GetImmunity"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "SquareLoseTurn"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "GetImmunity"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "SquareLoseTurn"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "GetImmunity"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "SquareLoseTurn"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "GetImmunity"},
-		{"SquareInfo": "Empty"},
-		{"SquareInfo": "SquareLoseTurn"},
-		{"SquareInfo": "Empty"},
+		{"SquareInfo": "loseTurn"},
 		{"SquareInfo": "Empty"},
 		{"SquareInfo": "Empty"},
 		{"SquareInfo": "Empty"},
 		{"SquareInfo": "GetImmunity"},
 		{"SquareInfo": "Empty"},
-		{"SquareInfo": "SquareLoseTurn"},
+		{"SquareInfo": "loseTurn"},
 		{"SquareInfo": "Empty"},
 		{"SquareInfo": "Empty"},
-		{"SquareInfo": "SquareLoseTurn"},
+		{"SquareInfo": "GetImmunity"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "loseTurn"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "GetImmunity"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "loseTurn"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "GetImmunity"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "loseTurn"},
 		{"SquareInfo": "Empty"},
 		{"SquareInfo": "Empty"},
 		{"SquareInfo": "Empty"},
 		{"SquareInfo": "Empty"},
 		{"SquareInfo": "GetImmunity"},
 		{"SquareInfo": "Empty"},
-		{"SquareInfo": "SquareLoseTurn"},
+		{"SquareInfo": "loseTurn"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "loseTurn"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "GetImmunity"},
+		{"SquareInfo": "Empty"},
+		{"SquareInfo": "loseTurn"},
 		{"SquareInfo": "Empty"},
 		{"SquareInfo": "Empty"}
 	]
